pyproject.toml
[project]
name = "anime-tagger"
version = "0.1.0"
description = "Classify and tag anime images using CLIP and WD-ViT-Tagger-v3"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "gradio>=4.44.0",
    "transformers>=4.40.0",
    "torch>=2.2.0",
    "onnxruntime>=1.17.0",
    "huggingface-hub>=0.22.0",
    "Pillow>=10.2.0",
    "numpy>=1.26.0",
]

[project.scripts]
anime-tagger = "anime_tagger.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["anime_tagger"]

[tool.uv]
package = true
# Use PyTorch CPU-only wheels (≈200 MB instead of ≈2.7 GB CUDA build).
# To switch to CUDA 12.1: change torch-backend to "cu121"
torch-backend = "cpu"
